message_store: Migrate message_store to typescript.
